refactor: filtered products fallback (#6)
