+++ src/widgets/product/index.ts
@@ -1,6 +1,8 @@
 export { BestProducts } from './ui/BestProducts';
 export { RecentProducts } from './ui/RecentProducts';
 export { BrandProducts } from './ui/BrandProducts';
-export { FilteredProducts } from './ui/FilteredProducts';
+export { FilteredSection } from './ui/FilteredSection';
 export { ProductLikes } from './ui/ProductLikes';
 export { ProductFilter } from './ui/ProductFilter';
+
+export { FilteredProductsSkeleton } from '#widgets/product/ui/FilteredProducts/FilteredProductsSkeleton';
